fix draw textture 2D vs 3D switching; update testing
