+++ packages/niivue/package.json
@@ -78,7 +78,7 @@
     "@types/ungap__structured-clone": "^1.2.0",
     "@typescript-eslint/eslint-plugin": "^7.16.1",
     "@typescript-eslint/parser": "^7.16.1",
-    "@vitest/coverage-v8": "^2.0.3",
+    "@vitest/coverage-v8": "^3.1.3",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-config-standard": "^17.1.0",
@@ -94,7 +94,7 @@
     "typedoc": "^0.26.4",
     "typescript": "^5.5.3",
     "vite": "^5.3.4",
-    "vitest": "^2.0.3"
+    "vitest": "^3.1.3"
   },
   "optionalDependencies": {
     "@rollup/rollup-linux-x64-gnu": "^4.18.1"
