+++ packages/niivue/src/index_histo.html
@@ -73,7 +73,7 @@
     <footer id="statusText">&nbsp;</footer>
     <script type="module" async>
       import { Niivue, NVImage, NVMesh, NVMeshLoaders, SHOW_RENDER, DRAG_MODE } from './niivue/index.ts'
-      drawSelect.onchange = async function () {
+      drawSelect.onchange = function () {
         const mode = parseInt(this.value)
         nv1.setDrawingEnabled(mode >= 0)
         if (mode >= 0) nv1.setPenValue(mode & 7, mode > 7)
@@ -81,7 +81,7 @@
           //erase selected cluster
           nv1.setPenValue(-0)
       }
-      opacitySlider.onchange = async function () {
+      opacitySlider.onchange = function () {
         nv1.setDrawOpacity(this.value * 0.01)
       }
       formatSelect.onchange = async function () {
@@ -151,20 +151,20 @@
         ]
         await nv1.loadVolumes(volumeLists[this.selectedIndex])
       }
-      aboutBtn.onclick = async function () {
+      aboutBtn.onclick = function () {
         alert(nv1.volumes[0].hdr.toFormattedString())
         // await nv1.loadDrawingFromUrl("./retina_draw.nii.gz")
       }
-      undoBtn.onclick = async function () {
+      undoBtn.onclick = function () {
         nv1.drawUndo()
       }
       saveBtn.onclick = async function () {
-        nv1.saveImage({ filename: 'test.nii', isSaveDrawing: true})
+        await nv1.saveImage({ filename: 'test.nii', isSaveDrawing: true})
       }
-      prevBtn.onclick = async function () {
+      prevBtn.onclick = function () {
          nv1.moveCrosshairInVox(0, 0, -1)
       }
-      nextBtn.onclick = async function () {
+      nextBtn.onclick = function () {
          nv1.moveCrosshairInVox(0, 0, 1)
       }
       dragSelect.onchange = function () {
@@ -208,6 +208,12 @@
         if (nv1.volumes[0].hdr.dims[3] > 1) {
           prevBtn.style.display = 'inline-block'
           nextBtn.style.display = 'inline-block'
+          // if dims[1, 2, 3] > 2048 set nv1.opts.is2DSliceShader = true
+          if (nv1.volumes[0].hdr.dims[1] > 2048 || nv1.volumes[0].hdr.dims[2] > 2048 || nv1.volumes[0].hdr.dims[3] > 2048) {
+            nv1.opts.is2DSliceShader = true
+          } else {
+            nv1.opts.is2DSliceShader = false
+          }
           if (nv1.opts.is2DSliceShader) {
             nv1.setSliceType(nv1.sliceTypeAxial)
             nv1.setVolumeRenderIllumination(-1)
