+++ packages/niivue/tests/unit/nvimage.test.ts
@@ -33,15 +33,3 @@ test('nvimage clone has unique ID', async () => {
   const clone = image.clone()
   expect(clone.id).not.toBe(image.id)
 })
-
-test('nvimage getValue', async () => {
-  const name = 'hippo.nii.gz'
-  const dataBuffer = readFileSync(path.join('./tests/images/', name))
-  const image = await NVImage.new(dataBuffer.buffer, name)
-  const voxX = 31
-  const voxY = 54
-  const voxZ = 27
-  const value = image.getValue(voxX, voxY, voxZ)
-  // The expected value at this position should be 9
-  expect(value).toBe(9)
-})
