+++ packages/niivue/vitest.config.ts
@@ -3,6 +3,7 @@ import { defineConfig, coverageConfigDefaults } from 'vitest/config'
 export default defineConfig({
   test: {
     fileParallelism: false,
+    maxConcurrency: 1,
     coverage: {
       provider: 'v8',
       reporter: ['text', 'html'],
